Allow markdown as a waka language ...begrudingly
There's a bug that prevents the target number of languages from being shown if one of them is ignored (or maybe it's specific to markdown, dunno). So fine, markdown can count as a language. I'm sure it won't be in the top 5 forever anyway.....right?

(I enjoy thorough documentation, so it probably will 😅)